Refactor LegendIcon into an exposed LegendIconWidget
